+++ lib/phoenix_live_view/channel.ex
@@ -5,7 +5,7 @@ defmodule Phoenix.LiveView.Channel do
   require Logger
 
   alias Phoenix.LiveView.{Socket, Utils, Diff, Upload, UploadConfig, Route, Session, Lifecycle}
-  alias Phoenix.Socket.Message
+  alias Phoenix.Socket.{Broadcast, Message}
 
   @prefix :phoenix
   @not_mounted_at_router :not_mounted_at_router
@@ -97,6 +97,11 @@ defmodule Phoenix.LiveView.Channel do
         |> view_handle_info(socket)
         |> handle_result({:handle_info, 2, nil}, state)
     end
+  end
+
+  def handle_info(%Broadcast{event: "phx_drain"}, state) do
+    send(state.socket.transport_pid, :socket_drain)
+    {:stop, {:shutdown, :draining}, state}
   end
 
   def handle_info(%Message{topic: topic, event: "phx_leave"} = msg, %{topic: topic} = state) do
@@ -1137,13 +1142,23 @@ defmodule Phoenix.LiveView.Channel do
   defp put_container(%Session{} = session, %Route{} = route, %{} = diff) do
     if container = session.redirected? && Route.container(route) do
       {tag, attrs} = container
-      Map.put(diff, :container, [tag, Enum.into(attrs, %{})])
+
+      attrs = attrs |> resolve_class_attribute_as_list() |> Enum.into(%{})
+
+      Map.put(diff, :container, [tag, attrs])
     else
       diff
     end
   end
 
   defp put_container(%Session{}, nil = _route, %{} = diff), do: diff
+
+  defp resolve_class_attribute_as_list(attrs) do
+    case attrs[:class] do
+      c when is_list(c) -> Keyword.put(attrs, :class, Enum.join(c, " "))
+      _ -> attrs
+    end
+  end
 
   defp reply_mount(result, from, %Session{} = session, route) do
     case result do